feat: group management, DingTalk forwarder, settings persistence
- Add AppSettings model and SettingsService for JSON persistence
- Fix Group ID type from int to long (supports large IDs like 51128424448184)
- Add DingTalkForwarder with markdown support, images, files, HMAC-SHA256 signing
- Enhance FeishuForwarder with rich post format, image/file link forwarding
- Rewrite SettingsWindow: 4 tabs (星球管理, 转发设置, 监控设置, 账号)
- Update MainWindow: URL input for adding groups, group add/remove in sidebar
- Update App.xaml.cs to create and inject SettingsService

+++ src/ZsxqForwarder.Core/Api/ZsxqApiClient.cs
@@ -62,7 +62,7 @@ public class ZsxqApiClient
     }
 
     public async Task<(List<Topic> Topics, bool IsEnd)> GetTopicsAsync(
-        int groupId, int count = 20, bool backward = true, long? endTime = null)
+        long groupId, int count = 20, bool backward = true, long? endTime = null)
     {
         var path = $"/v2/groups/{groupId}/topics";
         var businessParams = new Dictionary<string, string>
